feat: move Region normalization into consoles directly
To fix more complication regions we are moving away from region matching
inside of region.rs and moving more of the logic out to the individual
consoles.

This makes it easier to support regions like GB's "Non-Japan", where
using string reading is a worse approach (since it would match "Japan"
in that string and assume it's a Japan ROM.)

This avoids maintaining complicated logic and puts each console in
control of where the games are meant to be released.

This also normalizes the print output to our supported regions (instead
of seeing "GameGear International" you will see "USA, Europe, Asia" for
example.)